Move csv_path fixture to conftest.py
